+++ docs/guides/release-notes.md
@@ -13,9 +13,18 @@ Track all Zscaler SDK GO releases. New resources, features, and bug fixes will b
 
 ---
 
-``Last updated: v2.3.4``
+``Last updated: v2.31.4``
 
 ---
+
+# 2.31.4 (December 19, 2023)
+
+## Notes
+- Golang: **v1.19**
+
+### Enhacements
+
+- [PR #210](https://github.com/zscaler/zscaler-sdk-go/pull/210) - Added support to ZPA App Connector Controller Assistant Schedule `assistantSchedule`
 
 # 2.3.4 (December 19, 2023)